I fixed all the errors

+++ src/redux/reducer/ReducerController.js
@@ -6,12 +6,22 @@ export const initialState = {
 export const reducer = (state = initialState, action) => {
   switch (action.type) {
     case "FAVORITE_RECIPES":
-      const newFavorites = [...state.favoriteRecipes, action.recepes];
+      const recipeToAdd = action.recipe;
+
+      // Retsept allaqachon mavjudligini tekshirish
+      const exists = state.favoriteRecipes.some(recipe => recipe.id === recipeToAdd.id);
+      if (exists) {
+        return state;
+      }
+
+      // Agar mavjud bo'lmasa, uni favorite sahifasiga qo'shish
+      const newFavorites = [...state.favoriteRecipes, recipeToAdd];
       localStorage.setItem("favoriteRecipes", JSON.stringify(newFavorites));
       return {
         ...state,
         favoriteRecipes: newFavorites,
       };
+
     case "REMOVE_FAVORITE_RECIPE":
       const updatedFavorites = state.favoriteRecipes.filter(
         (recipe) => recipe.id !== action.recipeId
@@ -21,6 +31,7 @@ export const reducer = (state = initialState, action) => {
         ...state,
         favoriteRecipes: updatedFavorites,
       };
+
     default:
       return state;
   }
